renamed file and added comments
